More norm improvements

+++ src/parser/parser.c
@@ -33,11 +33,10 @@ int	add_camera(char **line, t_cam *camera)
 	if (line == NULL)
 		return (5);
 	camera->position = get_vector(line[1]);
-//	camera->position.y *= -1;
 	camera->normal = get_vector(line[2]);
 	vector_normalize(&(camera->normal));
 	camera->fov = ft_atoi(line[3]);
-	camera->aspect_ratio = RESOLUTION_X / RESOLUTION_Y;
+	camera->aspect_ratio = (float) RESOLUTION_X / RESOLUTION_Y;
 	ft_free_char_arr(line);
 	return (0);
 }
